feature: add authE to posts

+++ backend/src/main/java/com/revature/backend/posts/PostService.java
@@ -1,8 +1,11 @@
 package com.revature.backend.posts;
 
+import com.revature.backend.auth.InvalidCredentialsException;
 import com.revature.backend.users.User;
 import com.revature.backend.users.UserRepository;
+import com.revature.backend.utils.AuthenticationFacade;
 import com.revature.backend.utils.EntityNotFoundException;
+import com.revature.backend.utils.IAuthenticationFacade;
 import jakarta.validation.Valid;
 import lombok.Data;
 import org.springframework.stereotype.Service;
@@ -17,21 +20,33 @@ public class PostService {
 
 	private final PostRepository postRepository;
 	private final UserRepository userRepository;
+	private final IAuthenticationFacade authenticationFacade;
 
-	private void checkPostOwnership(Integer pathId) {
+	private void checkPostOwnership(Post post) {
 		/*
 		 * Check if the logged-in user can modify the blog post specified by the URL path.
 		 * TODO: add role-based check; if logged-in as admin, should be able to bypass ownership check.
 		 */
+		Integer claimedUserId = authenticationFacade.getClaimedUserId();
 
+		Optional<User> claimedUser = userRepository.findByIdAndDeletedFalse(claimedUserId);
+		if (claimedUser.isEmpty()) {
+			throw new InvalidCredentialsException(
+					String.format(
+							"Current user %s is deleted",
+							claimedUserId));
+		}
+
+		if (!claimedUserId.equals(post.getAuthor().getId())) {
+			throw new InvalidCredentialsException(claimedUserId, "post", post.getId());
+		}
 	}
 
 	public Post createBlogPost(@Valid PostDto postDto) {
-		// TODO: authorId shouldn't be part of the postDto, it should be obtained from user token/cookie.
-		Integer authorId = postDto.getAuthorId();
-		Optional<User> author = userRepository.findByIdAndDeletedFalse(authorId);
+		Integer claimedUserId = authenticationFacade.getClaimedUserId();
+		Optional<User> author = userRepository.findByIdAndDeletedFalse(claimedUserId);
 		if (author.isEmpty()) {
-			throw new EntityNotFoundException("user", authorId);
+			throw new EntityNotFoundException("user", claimedUserId);
 		}
 
 		Post post = new Post();
@@ -59,24 +74,27 @@ public class PostService {
 	}
 
 	public Post updateBlogPost(Integer postId, @Valid PostDto postDto) {
-		// TODO: AUTH: pull userId from token, etc
 		Optional<Post> post = postRepository.findById(postId);
 		if (post.isEmpty()) {
 			throw new EntityNotFoundException("post", postId);
-		} // else if author not found, or author is the wrong author, throw an exception
+		}
 
 		Post p = post.get();
+		checkPostOwnership(p);
+
 		p.setContent(postDto.getContent());
 		p.setUpdatedAt(LocalDateTime.now());
 		return postRepository.save(p);
 	}
 
 	public void deleteBlogPost(Integer postId) {
-		// TODO: auth, make sure user is [author of post] OR [admin role]
 		Optional<Post> post = postRepository.findById(postId);
 		if (post.isEmpty()) {
 			throw new EntityNotFoundException("post", postId);
 		}
+
+		Post p = post.get();
+		checkPostOwnership(p);
 
 		postRepository.deleteById(postId);
 	}
